Add files via upload
As we all come across ATMs in our daily lives, This project consists of five different classes, and it’s a console-based application. When the system starts, the user is prompted with a user id and a user pin. On entering the details successfully, the ATM functionalities are unlocked. The project allows us to perform the following operations.
1.	Transaction History
2.	Withdraw
3.	Deposit
4.	Transfer